Moves towards resource pipelines
