+++ harvester/core/processors/pipeline/resource.py
